fix models (add schema joi)
